refactor(plan): split commands/plan.js into one file per sub-command (audit rec #4)
Audit recommendation #4 (plan.js part): commands/plan.js had grown to
549 LOC with 8 sub-commands + 2 formatters in a single file. Split
into commands/plan/ with one file per sub-command, behaviour unchanged.

## Layout

  src/commands/plan.js                14 LOC (was 549, now thin re-export)
  src/commands/plan/index.js          14 LOC (barrel)
  src/commands/plan/_shared.js        40 LOC (formatPlan, formatHuTable)
  src/commands/plan/generate.js      284 LOC (planGenerateCommand + impl)
  src/commands/plan/list.js           21 LOC
  src/commands/plan/show.js           81 LOC
  src/commands/plan/ready.js          46 LOC
  src/commands/plan/validate.js       20 LOC
  src/commands/plan/delete.js          9 LOC
  src/commands/plan/add-hu.js         21 LOC
  src/commands/plan/remove-hu.js      17 LOC

## Strategy: thin re-export shim

`src/commands/plan.js` becomes `export * from "./plan/index.js";` so
the 11 external callers (cli.js + 9 test files) need no changes. The
canonical definitions live in plan/.

## Why generate.js is 284 LOC (over the 150 LOC suggestion)

`planGenerateImpl` is a single linear pipeline (planner → HUs build →
tests synthesizer → reviewer pass → save → render → board start).
Splitting further would scatter one transactional flow across files
and hurt readability. Kept intact intentionally.

Verified:
- node bin/kj.js plan --help              → 8 sub-commands listed
- node bin/kj.js plan {generate,show} --help → flag sets intact
- npx eslint src/                         → 0 errors
- npx vitest run                          → 4192/4192 passing
