+++ src/commands/plan/_shared.js
@@ -1,0 +1,40 @@
+// ---- Formatting helpers shared across plan sub-commands ----
+
+export function formatPlan(plan) {
+  const lines = [];
+  if (plan.approach) lines.push("## Approach", plan.approach, "");
+  if (plan.steps?.length) {
+    lines.push("## Steps");
+    for (let i = 0; i < plan.steps.length; i++) {
+      const step = plan.steps[i];
+      const commit = step.commit ? ` → \`${step.commit}\`` : "";
+      lines.push(`${i + 1}. ${step.description}${commit}`);
+    }
+    lines.push("");
+  }
+  if (plan.risks?.length) {
+    lines.push("## Risks");
+    for (const risk of plan.risks) lines.push(`- ${risk}`);
+    lines.push("");
+  }
+  if (plan.outOfScope?.length) {
+    lines.push("## Out of scope");
+    for (const item of plan.outOfScope) lines.push(`- ${item}`);
+    lines.push("");
+  }
+  return lines.join("\n");
+}
+
+export function formatHuTable(hus) {
+  if (!hus?.length) return "  (no HUs)";
+  const lines = [];
+  const maxTitle = Math.min(50, Math.max(...hus.map(h => h.title?.length || 0)));
+  lines.push(`  ${"ID".padEnd(30)} ${"Title".padEnd(maxTitle)} ${"Type".padEnd(10)} ${"Status".padEnd(10)} Deps`);
+  lines.push(`  ${"─".repeat(30)} ${"─".repeat(maxTitle)} ${"─".repeat(10)} ${"─".repeat(10)} ${"─".repeat(15)}`);
+  for (const hu of hus) {
+    const title = (hu.title || "").slice(0, maxTitle).padEnd(maxTitle);
+    const deps = (hu.blocked_by || []).join(", ") || "—";
+    lines.push(`  ${hu.id.padEnd(30)} ${title} ${(hu.task_type || "sw").padEnd(10)} ${(hu.status || "?").padEnd(10)} ${deps}`);
+  }
+  return lines.join("\n");
+}
